Fix input recorder
