+++ crates/maple_core/src/stdio_server/mod.rs
@@ -9,7 +9,6 @@ mod vim;
 
 pub use self::input::InputHistory;
 use self::input::{Event, ProviderEvent};
-pub use self::provider::SearchContext;
 use self::provider::{create_provider, Context};
 use self::rpc::{Call, MethodCall, Notification, RpcClient};
 use self::session::SessionManager;
